Add reply function

+++ src/main/java/com/ldh/forum/comment/service/CommentService.java
@@ -9,7 +9,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class CommentService {
@@ -24,16 +23,33 @@ public class CommentService {
         this.eventPublisher = eventPublisher;
     }
 
-    public List<Comment> getCommentsByBoardId(Long boardId) {
-        return commentRepository.findByBoardIdAndDeletedFalse(boardId);
+    public Comment addComment(Long boardId, String content, String author) {
+
+        Board board = boardRepository.findById(boardId).orElseThrow();
+        Comment comment = new Comment();
+        comment.setBoard(board);
+        comment.setContent(content);
+        comment.setAuthor(author);
+        return commentRepository.save(comment);
     }
 
-    public void addComment(Long boardId, String content, String author) {
-        Optional<Board> board = boardRepository.findById(boardId);
-        board.ifPresent(b -> {
-            Comment comment = new Comment(content, author, b);
-            commentRepository.save(comment);
-        });
+    public Comment addReply(Long parentId, String content, String author) {
+
+        Comment parentComment = commentRepository.findById(parentId).orElseThrow();
+        Comment reply = new Comment();
+        reply.setBoard(parentComment.getBoard());
+        reply.setContent(content);
+        reply.setAuthor(author);
+        reply.setParent(parentComment);
+        return commentRepository.save(reply);
+    }
+
+    public List<Comment> getCommentsByBoardId(Long boardId) {
+        return commentRepository.findByBoardIdAndParentIsNullOrderByCreatedAtAsc(boardId);
+    }
+
+    public List<Comment> getReplies(Long parentId) {
+        return commentRepository.findByParentIdOrderByCreatedAtAsc(parentId);
     }
 
     @Transactional
@@ -44,8 +60,13 @@ public class CommentService {
         if (!comment.getAuthor().equals(author))
             throw new SecurityException("You are not authorized to delete this comment.");
 
-        Board board = comment.getBoard();
+        if (!comment.getReplies().isEmpty()) {
+            comment.setContent("Deleted comment.");
+            commentRepository.save(comment);
+            return;
+        }
 
+        Board board = comment.getBoard();
         board.removeComment(comment);
 
         commentRepository.delete(comment);
